Rename experimental to fluent in tests.
