+++ tests/datasource/fluent/data_asset/data_connector/test_azure_blob_storage_data_connector.py
@@ -167,7 +167,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
     expected: List[BatchDefinition] = [
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_azure_blob_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -180,7 +180,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_azure_blob_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -193,7 +193,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_azure_blob_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -206,7 +206,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_azure_blob_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -219,7 +219,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_azure_blob_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -232,7 +232,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_azure_blob_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -245,7 +245,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_azure_blob_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -258,7 +258,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_azure_blob_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -271,7 +271,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_azure_blob_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -284,7 +284,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_azure_blob_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -360,7 +360,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #     expected: List[BatchDefinition] = [
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_azure_blob_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "alex_20200809_1000.csv", "name": "alex", "timestamp": "20200809", "price": "1000"}
@@ -368,7 +368,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_azure_blob_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "eugene_20200809_1500.csv", "name": "eugene", "timestamp": "20200809", "price": "1500"}
@@ -376,7 +376,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_azure_blob_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "james_20200811_1009.csv", "name": "james", "timestamp": "20200811", "price": "1009"}
@@ -384,7 +384,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_azure_blob_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "abe_20200809_1040.csv", "name": "abe", "timestamp": "20200809", "price": "1040"}
@@ -392,7 +392,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_azure_blob_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "will_20200809_1002.csv", "name": "will", "timestamp": "20200809", "price": "1002"}
@@ -400,7 +400,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_azure_blob_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "james_20200713_1567.csv", "name": "james", "timestamp": "20200713", "price": "1567"}
@@ -408,7 +408,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_azure_blob_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "eugene_20201129_1900.csv", "name": "eugene", "timestamp": "20201129", "price": "1900"}
@@ -416,7 +416,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_azure_blob_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "will_20200810_1001.csv", "name": "will", "timestamp": "20200810", "price": "1001"}
@@ -424,7 +424,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_azure_blob_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "james_20200810_1003.csv", "name": "james", "timestamp": "20200810", "price": "1003"}
@@ -432,7 +432,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_azure_blob_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "alex_20200819_1300.csv", "name": "alex", "timestamp": "20200819", "price": "1300"}
@@ -495,7 +495,7 @@ def test_return_only_unique_batch_definitions(mock_list_keys):
     expected: List[BatchDefinition] = [
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_azure_blob_storage_data_asset",
             batch_identifiers=IDDict(
                 {"path": "B/file_1.csv", "directory": "B", "filename": "file_1.csv"}
@@ -503,7 +503,7 @@ def test_return_only_unique_batch_definitions(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_azure_blob_storage_data_asset",
             batch_identifiers=IDDict(
                 {"path": "B/file_2.csv", "directory": "B", "filename": "file_2.csv"}
