+++ tests/datasource/fluent/data_asset/data_connector/test_google_cloud_storage_data_connector.py
@@ -166,7 +166,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
     expected: List[BatchDefinition] = [
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_google_cloud_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -179,7 +179,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_google_cloud_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -192,7 +192,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_google_cloud_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -205,7 +205,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_google_cloud_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -218,7 +218,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_google_cloud_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -231,7 +231,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_google_cloud_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -244,7 +244,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_google_cloud_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -257,7 +257,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_google_cloud_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -270,7 +270,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_google_cloud_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -283,7 +283,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_google_cloud_storage_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -358,7 +358,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #     expected: List[BatchDefinition] = [
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_google_cloud_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "alex_20200809_1000.csv", "name": "alex", "timestamp": "20200809", "price": "1000"}
@@ -366,7 +366,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_google_cloud_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "eugene_20200809_1500.csv", "name": "eugene", "timestamp": "20200809", "price": "1500"}
@@ -374,7 +374,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_google_cloud_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "james_20200811_1009.csv", "name": "james", "timestamp": "20200811", "price": "1009"}
@@ -382,7 +382,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_google_cloud_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "abe_20200809_1040.csv", "name": "abe", "timestamp": "20200809", "price": "1040"}
@@ -390,7 +390,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_google_cloud_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "will_20200809_1002.csv", "name": "will", "timestamp": "20200809", "price": "1002"}
@@ -398,7 +398,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_google_cloud_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "james_20200713_1567.csv", "name": "james", "timestamp": "20200713", "price": "1567"}
@@ -406,7 +406,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_google_cloud_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "eugene_20201129_1900.csv", "name": "eugene", "timestamp": "20201129", "price": "1900"}
@@ -414,7 +414,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_google_cloud_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "will_20200810_1001.csv", "name": "will", "timestamp": "20200810", "price": "1001"}
@@ -422,7 +422,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_google_cloud_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "james_20200810_1003.csv", "name": "james", "timestamp": "20200810", "price": "1003"}
@@ -430,7 +430,7 @@ def test_return_all_batch_definitions_unsorted(mock_list_keys):
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_google_cloud_storage_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"path": "alex_20200819_1300.csv", "name": "alex", "timestamp": "20200819", "price": "1300"}
@@ -492,7 +492,7 @@ def test_return_only_unique_batch_definitions(mock_list_keys):
     expected: List[BatchDefinition] = [
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_google_cloud_storage_data_asset",
             batch_identifiers=IDDict(
                 {"path": "B/file_1.csv", "directory": "B", "filename": "file_1.csv"}
@@ -500,7 +500,7 @@ def test_return_only_unique_batch_definitions(mock_list_keys):
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_google_cloud_storage_data_asset",
             batch_identifiers=IDDict(
                 {"path": "B/file_2.csv", "directory": "B", "filename": "file_2.csv"}
