+++ tests/datasource/fluent/data_asset/data_connector/test_s3_data_connector.py
@@ -183,7 +183,7 @@ def test_return_all_batch_definitions_unsorted():
     expected: List[BatchDefinition] = [
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_s3_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -196,7 +196,7 @@ def test_return_all_batch_definitions_unsorted():
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_s3_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -209,7 +209,7 @@ def test_return_all_batch_definitions_unsorted():
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_s3_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -222,7 +222,7 @@ def test_return_all_batch_definitions_unsorted():
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_s3_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -235,7 +235,7 @@ def test_return_all_batch_definitions_unsorted():
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_s3_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -248,7 +248,7 @@ def test_return_all_batch_definitions_unsorted():
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_s3_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -261,7 +261,7 @@ def test_return_all_batch_definitions_unsorted():
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_s3_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -274,7 +274,7 @@ def test_return_all_batch_definitions_unsorted():
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_s3_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -287,7 +287,7 @@ def test_return_all_batch_definitions_unsorted():
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_s3_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -300,7 +300,7 @@ def test_return_all_batch_definitions_unsorted():
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_s3_data_asset",
             batch_identifiers=IDDict(
                 {
@@ -393,7 +393,7 @@ def test_return_all_batch_definitions_unsorted():
 #     expected: List[BatchDefinition] = [
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_s3_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"name": "abe", "timestamp": "20200809", "price": "1040"}
@@ -401,7 +401,7 @@ def test_return_all_batch_definitions_unsorted():
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_s3_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"name": "alex", "timestamp": "20200819", "price": "1300"}
@@ -409,7 +409,7 @@ def test_return_all_batch_definitions_unsorted():
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_s3_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"name": "alex", "timestamp": "20200809", "price": "1000"}
@@ -417,7 +417,7 @@ def test_return_all_batch_definitions_unsorted():
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_s3_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"name": "eugene", "timestamp": "20201129", "price": "1900"}
@@ -425,7 +425,7 @@ def test_return_all_batch_definitions_unsorted():
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_s3_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"name": "eugene", "timestamp": "20200809", "price": "1500"}
@@ -433,7 +433,7 @@ def test_return_all_batch_definitions_unsorted():
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_s3_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"name": "james", "timestamp": "20200811", "price": "1009"}
@@ -441,7 +441,7 @@ def test_return_all_batch_definitions_unsorted():
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_s3_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"name": "james", "timestamp": "20200810", "price": "1003"}
@@ -449,7 +449,7 @@ def test_return_all_batch_definitions_unsorted():
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_s3_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"name": "james", "timestamp": "20200713", "price": "1567"}
@@ -457,7 +457,7 @@ def test_return_all_batch_definitions_unsorted():
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_s3_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"name": "will", "timestamp": "20200810", "price": "1001"}
@@ -465,7 +465,7 @@ def test_return_all_batch_definitions_unsorted():
 #         ),
 #         BatchDefinition(
 #             datasource_name="my_file_path_datasource",
-#             data_connector_name="experimental",
+#             data_connector_name="fluent",
 #             data_asset_name="my_s3_data_asset",
 #             batch_identifiers=IDDict(
 #                 {"name": "will", "timestamp": "20200809", "price": "1002"}
@@ -571,7 +571,7 @@ def test_return_only_unique_batch_definitions():
     expected: List[BatchDefinition] = [
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_s3_data_asset",
             batch_identifiers=IDDict(
                 {"path": "B/file_1.csv", "directory": "B", "filename": "file_1.csv"}
@@ -579,7 +579,7 @@ def test_return_only_unique_batch_definitions():
         ),
         BatchDefinition(
             datasource_name="my_file_path_datasource",
-            data_connector_name="experimental",
+            data_connector_name="fluent",
             data_asset_name="my_s3_data_asset",
             batch_identifiers=IDDict(
                 {"path": "B/file_2.csv", "directory": "B", "filename": "file_2.csv"}
@@ -855,7 +855,7 @@ def test_foxtrot():
 #             instantiate_class_from_config(
 #                 config=my_data_connector_yaml,
 #                 runtime_environment={
-#                     "name": "experimental",
+#                     "name": "fluent",
 #                     "execution_engine": PandasExecutionEngine(),
 #                 },
 #                 config_defaults={
