+++ tests/datasource/fluent/test_sqlite_datasource.py
@@ -9,7 +9,7 @@ from pydantic import ValidationError
 
 from great_expectations.datasource.fluent import SqliteDatasource
 from great_expectations.datasource.fluent.sqlite_datasource import SqliteDsn
-from tests.experimental.datasources.conftest import sqlachemy_execution_engine_mock_cls
+from tests.datasource.fluent.conftest import sqlachemy_execution_engine_mock_cls
 
 
 @pytest.fixture
